# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "OpenApi description",
        "contact": {
            "email": "support@mysite.com"
        },
        "license": {
            "name": "Proprietary",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:7501/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Member Authentication"
                ],
                "summary": "Register a new member",
                "description": "Registers a new member account and returns basic profile info and the activation code.",
                "operationId": "ff4a51db1f5af87391e3d07c02c9aff4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "nickname"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "yourPassword123"
                                    },
                                    "nickname": {
                                        "type": "string",
                                        "example": "JohnDoe"
                                    },
                                    "region_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uid": {
                                            "type": "integer",
                                            "example": 156490
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "nickname": {
                                            "type": "string",
                                            "example": "JohnDoe"
                                        },
                                        "activation_code": {
                                            "type": "string",
                                            "example": "A1B2C3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/auth/activation": {
            "post": {
                "tags": [
                    "Member Authentication"
                ],
                "summary": "Activate member account",
                "description": "Activates a member account using the activation code.",
                "operationId": "9185c8b41bc65405ab0e451b05549617",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "activation_code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "activation_code": {
                                        "type": "string",
                                        "example": "A1B2C3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Account successfully activated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Account activation has been activated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Member Authentication"
                ],
                "summary": "Member login",
                "description": "Authenticates a member and returns a JWT token.",
                "operationId": "1d57db395a65b205bca3bd93a45b89ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "yourPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGci..."
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "500": {
                        "description": "Could not create token"
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Member Authentication"
                ],
                "summary": "Refresh JWT token",
                "description": "Refreshes the JWT token for the authenticated user.",
                "operationId": "a18851f4a9969e3b031c98c7f609c808",
                "responses": {
                    "202": {
                        "description": "Token refreshed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGci..."
                                        },
                                        "token_expired": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGci..."
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token not registered"
                    },
                    "422": {
                        "description": "Token cannot be refreshed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Member Authentication"
                ],
                "summary": "Logout member",
                "description": "Terminates the current JWT token and logs out the member.",
                "operationId": "f62f1b6af10bdc562cdd99e2d1e31980",
                "responses": {
                    "202": {
                        "description": "Token terminated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token_expired": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGci..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Token not registered"
                    },
                    "304": {
                        "description": "Token is already terminated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/whoami": {
            "get": {
                "tags": [
                    "Member Authentication"
                ],
                "summary": "Get authenticated user info",
                "description": "Returns information about the authenticated user.",
                "operationId": "ed1772df8d000912d8aff80eb813eb4f",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "uid": {
                                            "type": "integer",
                                            "example": 156490
                                        },
                                        "nickname": {
                                            "type": "string",
                                            "example": "JohnDoe"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "http://..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Member Authentication",
            "description": "Endpoints about the authenticated user"
        }
    ]
}